Задание 5.1
